requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyJWT
passlib[bcrypt]
python-dotenv
PyMySQL
